Use float instead of double as output

+++ rtkprojections.cxx
@@ -11,7 +11,7 @@ int main(int argc, char * argv[])
 {
   GGO(rtkprojections, args_info);
 
-  typedef double OutputPixelType;
+  typedef float OutputPixelType;
   const unsigned int Dimension = 3;
   typedef itk::Image< OutputPixelType, Dimension > OutputImageType;
